# Optimize MIR type data structure