--- a/lib/mir/src/types/record.rs
+++ b/lib/mir/src/types/record.rs
@@ -1,11 +1,15 @@
+use std::sync::Arc;
+
 #[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Record {
-    name: String,
+    name: Arc<str>,
 }
 
 impl Record {
     pub fn new(name: impl Into<String>) -> Self {
-        Self { name: name.into() }
+        Self {
+            name: name.into().into(),
+        }
     }
 
     pub fn name(&self) -> &str {