… lib/hir-mir/src/snapshots/hir_mir__built_in_call__tests__compile_debug.snap